OBPIH-6500 Move Invoice options form and items table into a separate components
