+++ src/js/components/invoice/create/ConfirmInvoicePage.jsx
@@ -5,218 +5,14 @@ import _ from 'lodash';
 import PropTypes from 'prop-types';
 import { Form } from 'react-final-form';
 import { connect } from 'react-redux';
-import { Tooltip } from 'react-tippy';
 
 import { hideSpinner, showSpinner } from 'actions';
 import invoiceApi from 'api/services/InvoiceApi';
-import DocumentButton from 'components/DocumentButton';
-import ArrayField from 'components/form-elements/ArrayField';
-import DateField from 'components/form-elements/DateField';
-import LabelField from 'components/form-elements/LabelField';
-import TextField from 'components/form-elements/TextField';
-import { INVOICE_URL, ORDER_URL, STOCK_MOVEMENT_URL } from 'consts/applicationUrls';
-import { renderFormField } from 'utils/form-utils';
-import { getInvoiceDescription } from 'utils/form-values-utils';
+import InvoiceItemsTable from 'components/invoice/create/InvoiceItemsTable';
+import InvoiceOptionsForm from 'components/invoice/create/InvoiceOptionsForm';
+import { INVOICE_URL } from 'consts/applicationUrls';
 import accountingFormat from 'utils/number-utils';
 import Translate from 'utils/Translate';
-
-const INVOICE_HEADER_FIELDS = {
-  invoiceNumber: {
-    type: TextField,
-    label: 'react.invoice.invoiceNumber.label',
-    defaultMessage: 'Invoice number',
-    attributes: {
-      disabled: true,
-    },
-  },
-  vendorName: {
-    type: TextField,
-    label: 'react.Invoice.vendor.label',
-    defaultMessage: 'Vendor',
-    attributes: {
-      disabled: true,
-    },
-  },
-  vendorInvoiceNumber: {
-    type: TextField,
-    label: 'react.invoice.vendorInvoiceNumber.label',
-    defaultMessage: 'Vendor Invoice Number',
-    attributes: {
-      disabled: true,
-    },
-  },
-  dateInvoiced: {
-    type: DateField,
-    label: 'react.invoice.invoiceDate.label',
-    defaultMessage: 'Invoice Date',
-    attributes: {
-      disabled: true,
-      dateFormat: 'MM/DD/YYYY',
-    },
-  },
-  'currencyUom.code': {
-    type: TextField,
-    label: 'react.invoice.currency.label',
-    defaultMessage: 'Currency',
-    attributes: {
-      disabled: true,
-    },
-  },
-  totalValue: {
-    type: TextField,
-    label: 'react.invoice.total.label',
-    defaultMessage: 'Total',
-    attributes: {
-      disabled: true,
-    },
-    getDynamicAttr: ({ values }) => ({
-      className: values && values.totalValue && (values.totalValue < 0 || values.totalValue.startsWith('(')) ? 'negative-value' : '',
-    }),
-  },
-};
-
-const INVOICE_ITEMS = {
-  invoiceItems: {
-    type: ArrayField,
-    virtualized: true,
-    totalCount: ({ totalCount }) => totalCount,
-    isRowLoaded: ({ isRowLoaded }) => isRowLoaded,
-    loadMoreRows: ({ loadMoreRows }) => loadMoreRows(),
-    isFirstPageLoaded: ({ isFirstPageLoaded }) => isFirstPageLoaded,
-    getDynamicRowAttr: ({ rowValues }) => ({ className: rowValues && rowValues.totalAmount && rowValues.totalAmount < 0 ? 'negative-row-value' : '' }),
-    fields: {
-      prepaymentIcon: {
-        type: (params) => {
-          const { values } = params;
-          const hasItems = values && values.invoiceItems;
-          const isPrepLine = hasItems && (values.isPrepaymentInvoice ||
-              values.invoiceItems[params.rowIndex].isPrepaymentItem);
-          if (isPrepLine) {
-            return (
-              <div className="d-flex align-items-center justify-content-center">
-                <Tooltip
-                  html="Prepayment"
-                  theme="transparent"
-                  delay="150"
-                  duration="250"
-                  hideDelay="50"
-                >
-                  {/* &#x24C5; = hexadecimal circled letter P */}
-                  <b>&#x24C5;</b>
-                </Tooltip>
-              </div>
-            );
-          }
-          return null;
-        },
-        label: '',
-        defaultMessage: '',
-        flexWidth: '0.25',
-      },
-      orderNumber: {
-        type: LabelField,
-        label: 'react.invoice.orderNumber.label',
-        defaultMessage: 'PO Number',
-        flexWidth: '1',
-        getDynamicAttr: ({ values, rowIndex }) => {
-          const orderId = values && values.invoiceItems
-              && values.invoiceItems[rowIndex]
-              && values.invoiceItems[rowIndex].orderId;
-          return { url: orderId ? ORDER_URL.show(orderId): '' };
-        },
-      },
-      shipmentNumber: {
-        type: LabelField,
-        label: 'react.invoice.shipmentNumber.label',
-        defaultMessage: 'Shipment Number',
-        flexWidth: '1',
-        getDynamicAttr: ({ values, rowIndex }) => {
-          const shipmentId = values && values.invoiceItems
-              && values.invoiceItems[rowIndex]
-              && values.invoiceItems[rowIndex].shipmentId;
-          return { url: shipmentId ? STOCK_MOVEMENT_URL.show(shipmentId) : '' };
-        },
-      },
-      budgetCode: {
-        type: LabelField,
-        label: 'react.invoice.budgetCode.label',
-        defaultMessage: 'Budget Code',
-        flexWidth: '1',
-      },
-      glCode: {
-        type: LabelField,
-        label: 'react.invoice.glCode.label',
-        defaultMessage: 'GL Code',
-        flexWidth: '1',
-      },
-      productCode: {
-        type: LabelField,
-        label: 'react.invoice.itemNo.label',
-        defaultMessage: 'Item No',
-        flexWidth: '1',
-      },
-      description: {
-        type: LabelField,
-        label: 'react.invoice.description.label',
-        defaultMessage: 'Description',
-        flexWidth: '5',
-        attributes: {
-          className: 'text-left',
-        },
-        getDynamicAttr: params => ({
-          formatValue: () => {
-            const { values, rowIndex } = params;
-            const rowValue = values?.invoiceItems?.[rowIndex];
-            // If it's not an adjustment, but product, and it has a synonym, display it
-            // with a tooltip with the original name of the product
-            return getInvoiceDescription(rowValue);
-          },
-        }),
-      },
-      quantity: {
-        type: LabelField,
-        label: 'react.invoice.qty.label',
-        defaultMessage: 'Qty',
-        flexWidth: '1',
-        getDynamicAttr: params => ({
-          formatValue: () => {
-            const { values } = params;
-            const hasItems = values && values.invoiceItems;
-            const isPrepLine = hasItems && values.invoiceItems[params.rowIndex].isPrepaymentItem;
-            if (isPrepLine) {
-              return params.fieldValue * (-1);
-            }
-            return params.fieldValue;
-          },
-        }),
-      },
-      uom: {
-        type: LabelField,
-        label: 'react.invoice.uom.label',
-        defaultMessage: 'UOM',
-        flexWidth: '1',
-      },
-      unitPrice: {
-        type: LabelField,
-        label: 'react.invoice.unitPrice.label',
-        defaultMessage: 'Unit Price',
-        flexWidth: '1',
-        attributes: {
-          formatValue: value => (value ? accountingFormat(value) : value),
-        },
-      },
-      totalAmount: {
-        type: LabelField,
-        label: 'react.invoice.totalPrice.label',
-        defaultMessage: 'Total Price',
-        flexWidth: '1',
-        attributes: {
-          formatValue: value => (value ? accountingFormat(value) : value),
-        },
-      },
-    },
-  },
-};
 
 const PREPAYMENT_INVOICE = 'PREPAYMENT_INVOICE';
 
@@ -225,12 +21,9 @@ class ConfirmInvoicePage extends Component {
     super(props);
     this.state = {
       values: { ...this.props.initialValues, invoiceItems: [], totalValue: 0 },
-      isFirstPageLoaded: false,
     };
 
-    this.isRowLoaded = this.isRowLoaded.bind(this);
     this.loadMoreRows = this.loadMoreRows.bind(this);
-    this.toggleDropdown = this.toggleDropdown.bind(this);
   }
 
   componentDidMount() {
@@ -289,23 +82,11 @@ class ConfirmInvoicePage extends Component {
   }
 
   /**
-   * Checks if row is loaded, needed for pagination
-   * @param {string} index
-   * @public
-   */
-  isRowLoaded({ index }) {
-    return !!this.state.values.invoiceItems[index];
-  }
-
-  /**
    * Loads more rows, needed for pagination
    * @param {index} startIndex
    * @public
    */
   loadMoreRows({ startIndex }) {
-    this.setState({
-      isFirstPageLoaded: true,
-    });
     invoiceApi.getInvoiceItems(this.state.values.id, {
       params: { offset: startIndex, max: this.props.pageSize },
     })
@@ -358,15 +139,6 @@ class ConfirmInvoicePage extends Component {
       });
   }
 
-  /**
-   * Toggle the downloadable files
-   */
-  toggleDropdown() {
-    this.setState({
-      isDropdownVisible: !this.state.isDropdownVisible,
-    });
-  }
-
   render() {
     return (
       <div>
@@ -377,61 +149,13 @@ class ConfirmInvoicePage extends Component {
           mutators={{ ...arrayMutators }}
           render={({ handleSubmit, values }) => (
             <form onSubmit={handleSubmit}>
-              <div className="classic-form classic-form-condensed">
-                <span className="buttons-container classic-form-buttons">
-                  <button
-                    type="button"
-                    className="btn btn-outline-secondary float-right btn-form btn-xs"
-                    onClick={() => { window.location = INVOICE_URL.show(values.id); }}
-                  >
-                    <span><i className="fa fa-sign-out pr-2" /><Translate id="react.default.button.saveAndExit.label" defaultMessage="Save and exit" /></span>
-                  </button>
-                  <span className="mr-3">
-                    <div className="dropdown">
-                      <button
-                        type="button"
-                        onClick={this.toggleDropdown}
-                        className="dropdown-button float-right mb-1 btn btn-outline-secondary align-self-end btn-xs"
-                      >
-                        <span><i className="fa fa-sign-out pr-2" /><Translate id="react.default.button.download.label" defaultMessage="Download" /></span>
-                      </button>
-                      <div className={`dropdown-content print-buttons-container col-md-3 flex-grow-1
-                        ${this.state.isDropdownVisible ? 'visible' : ''}`}
-                      >
-                        {this.state.values.documents && this.state.values.documents.length > 0 &&
-                        _.map(this.state.values.documents, (document, idx) => {
-                          if (document.hidden) {
-                            return null;
-                          }
-                          return (<DocumentButton
-                            link={document.link}
-                            buttonTitle={document.name}
-                            {...document}
-                            key={idx}
-                            disabled={false}
-                          />);
-                        })}
-                      </div>
-                    </div>
-                  </span>
-                </span>
-                <div className="form-title"><Translate id="react.invoice.options.label" defaultMessage="Invoice options" /></div>
-                {_.map(INVOICE_HEADER_FIELDS, (fieldConfig, fieldName) =>
-                  renderFormField(fieldConfig, fieldName, {
-                    values,
-                    totalCount: this.state.values.totalCount,
-                    loadMoreRows: this.loadMoreRows,
-                    isRowLoaded: this.isRowLoaded,
-                    isFirstPageLoaded: this.state.isFirstPageLoaded,
-                  }))}
-              </div>
+              <InvoiceOptionsForm values={values} />
               <div className="submit-buttons">
                 <button
                   className="btn btn-outline-primary btn-form btn-xs"
-                  onClick={() => this.props.previousPage(this.state.values)}
-                  disabled={this.state.values.datePosted ||
-                    this.state.values.invoiceType === PREPAYMENT_INVOICE ||
-                    this.state.values.hasPrepaymentInvoice}
+                  onClick={() => this.props.previousPage(values)}
+                  disabled={values.datePosted
+                    || values.invoiceType === PREPAYMENT_INVOICE || values.hasPrepaymentInvoice}
                 >
                   <Translate id="react.default.button.previous.label" defaultMessage="Previous" />
                 </button>
@@ -439,7 +163,7 @@ class ConfirmInvoicePage extends Component {
                   type="submit"
                   onClick={() => { this.submitInvoice(); }}
                   className="btn btn-outline-success float-right btn-form btn-xs"
-                  disabled={this.state.values.dateSubmitted || this.state.values.datePosted}
+                  disabled={values.dateSubmitted || values.datePosted}
                 >
                   <Translate id="react.invoice.submit.label" defaultMessage="Submit for Approval" />
                 </button>
@@ -448,22 +172,15 @@ class ConfirmInvoicePage extends Component {
                     type="submit"
                     onClick={() => { this.postInvoice(); }}
                     className="btn btn-outline-success float-right btn-form btn-xs"
-                    disabled={this.state.values.datePosted}
+                    disabled={values.datePosted}
                   >
                     <Translate id="react.invoice.post.label" defaultMessage="Post Invoice" />
                   </button>}
               </div>
-              <div className="my-2 table-form">
-                {_.map(INVOICE_ITEMS, (fieldConfig, fieldName) =>
-                  renderFormField(fieldConfig, fieldName, {
-                    values: this.state.values,
-                    totalCount: this.state.values.totalCount,
-                    loadMoreRows: this.loadMoreRows,
-                    isRowLoaded: this.isRowLoaded,
-                    isFirstPageLoaded: this.state.isFirstPageLoaded,
-                    updateTotalCount: this.updateTotalCount,
-                  }))}
-              </div>
+              <InvoiceItemsTable
+                values={values}
+                loadMoreRows={this.loadMoreRows}
+              />
             </form>
           )}
         />
@@ -472,7 +189,7 @@ class ConfirmInvoicePage extends Component {
   }
 }
 
-const mapStateToProps = state => ({
+const mapStateToProps = (state) => ({
   pageSize: state.session.pageSize,
   isSuperuser: state.session.isSuperuser,
 });
